requirements.txt
Faker==40.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
python-dotenv==1.2.1
tqdm==4.67.1
